requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.6
decorator==5.1.1
executing==2.1.0
fastapi==0.115.5
google-auth==2.36.0
h11==0.14.0
idna==3.10
jedi==0.19.1
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
python-dotenv
gunicorn
google-auth
google-auth-oauthlib
google-api-python-client  
google-cloud-storage
